chore: merge bean factories (#66137)
